# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: FeeDLooP API
  description: |
    FeeDLooP is a feedback collection service for software development companies. It provides APIs for managing projects, collecting feedback through embeddable widgets, and managing bug reports, initiatives, and feedback submissions.

    ## Authentication
    Most endpoints require authentication using NextAuth.js session cookies. The widget submission endpoint is public to allow embedding.

    ## Rate Limiting
    API endpoints implement rate limiting to prevent abuse. Refer to security documentation for specific limits.

    ## CORS
    Different CORS policies are applied based on endpoint:
    - Widget endpoints: Permissive CORS for embedding
    - Dashboard/API endpoints: Restrictive CORS for security

    ## Widget Integration
    The embeddable widget uses the `/api/widget/submit` endpoint to submit feedback from any website. See the Framework Integration Guide for implementation details.

  version: 1.0.0
  contact:
    name: FeeDLooP API Support
    url: https://github.com/your-org/feedloop
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:3000
    description: Development server
  - url: https://localhost:3001
    description: Alternative development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/widget/submit:
    post:
      tags:
        - Widget
      summary: Submit feedback from widget
      description: |
        Accepts feedback submissions from the embeddable JavaScript widget. This endpoint supports multipart form data with file attachments and compressed diagnostic data.

        **Features:**
        - Accepts bug reports, initiatives, and feedback
        - Supports up to 5 file attachments (max 5MB each)
        - Handles compressed diagnostic data (gzip)
        - Validates project key
        - Public endpoint (no authentication required)

        **Security:**
        - Input validation and sanitization
        - File type validation
        - CORS headers for embedding
        - Rate limiting protection
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - project_key
                - type
                - title
                - description
              properties:
                project_key:
                  type: string
                  pattern: '^[a-zA-Z0-9]{16,32}$'
                  description: Unique project integration key
                  example: "test-project-key-12345678"
                type:
                  type: string
                  enum: [bug, initiative, feedback]
                  description: Type of feedback being submitted
                  example: "bug"
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Brief title describing the feedback
                  example: "Page loading issue on checkout"
                description:
                  type: string
                  minLength: 1
                  maxLength: 5000
                  description: Detailed description of the feedback
                  example: "The checkout page takes too long to load on mobile devices"
                reporter_name:
                  type: string
                  maxLength: 100
                  description: Optional name of the person submitting feedback
                  example: "John Doe"
                reporter_email:
                  type: string
                  format: email
                  maxLength: 255
                  description: Optional email of the person submitting feedback
                  example: "john@example.com"
                priority:
                  type: string
                  enum: [low, medium, high, critical]
                  description: Priority level (defaults to medium)
                  example: "medium"
                url:
                  type: string
                  format: uri
                  maxLength: 2000
                  description: URL where the feedback was submitted from
                  example: "https://example.com/checkout"
                user_agent:
                  type: string
                  maxLength: 500
                  description: Browser user agent string
                  example: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
                console_logs:
                  type: string
                  description: JSON string of console errors and logs
                network_requests:
                  type: string
                  description: JSON string of network request data
                diagnostic_data:
                  type: string
                  description: Uncompressed diagnostic data as JSON string
                diagnostic_data_compressed:
                  type: string
                  description: Base64 encoded gzip compressed diagnostic data
                compression_type:
                  type: string
                  enum: [gzip]
                  description: Compression algorithm used
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
                  description: File attachments (max 5 files, 5MB each)
            encoding:
              attachments:
                contentType: image/*, application/pdf, text/*
      responses:
        '201':
          description: Feedback submitted successfully
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: CORS header for widget embedding
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Feedback submitted successfully"
                  report_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  attachments_uploaded:
                    type: integer
                    example: 2
                  total_attachments:
                    type: integer
                    example: 3
                  attachments:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttachmentResponse'
                  attachment_errors:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                        error:
                          type: string
        '400':
          description: Validation failed or invalid project key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: Validation error
                  value:
                    error: "Validation failed"
                    details:
                      - field: "title"
                        message: "Title is required"
                invalid_project:
                  summary: Invalid project key
                  value:
                    error: "Invalid project key"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      tags:
        - Widget
      summary: CORS preflight for widget submissions
      description: Handles CORS preflight requests for widget embedding
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: Allowed origins for CORS
            Access-Control-Allow-Methods:
              schema:
                type: string
              description: Allowed HTTP methods
            Access-Control-Allow-Headers:
              schema:
                type: string
              description: Allowed headers

  # Project Management
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticates a user with email and password. Returns session cookie.
        This endpoint is typically handled by NextAuth.js.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    $ref: '#/components/schemas/User'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health Check
...
  /api/test-db:
    get:
      tags:
        - Health
      summary: Test database connection
      description: Tests database connectivity (development only)
      responses:
        '200':
          description: Database connection successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "connected"
                  database:
                    type: string
                    example: "postgresql"
        '500':
          description: Database connection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/test-verify-user:
    get:
      tags:
        - Health
      summary: Test user verification
      description: Tests user authentication and session verification (development only)
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "verified"
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    AttachmentResponse:
      type: object
      required:
        - id
        - filename
        - content_type
        - size
        - url
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique attachment identifier
        filename:
          type: string
          minLength: 1
          maxLength: 255
          description: Sanitized filename
        original_filename:
          type: string
          description: Original filename before sanitization
        content_type:
          type: string
          description: MIME type
        size:
          type: integer
          minimum: 1
          maximum: 5242880
          description: File size in bytes (max 5MB)
        url:
          type: string
          format: uri
          description: Download URL
        created_at:
          type: string
          format: date-time
          description: Upload timestamp
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: User identifier
        email:
          type: string
          format: email
          description: User email address
        name:
          type: string
          description: Full name
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Profile picture URL
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        last_login:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Specific error message
          description: Detailed validation errors
        code:
          type: string
          description: Error code for programmatic handling
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
security:
  - sessionAuth: []
...